--- workspaceEmpresa-03.6.5/projeto-empresav1/src/test/java/br/com/contmatic/empresav1/ITtest/DepartamentoIT.java
@@ -1,11 +1,0 @@
-package br.com.contmatic.empresav1.ITtest;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
-@RunWith(Suite.class)
-@SuiteClasses({})
-public class DepartamentoIT {
-
-}
