+++ workspaceEmpresa-03.7.9/projeto-empresav1/src/test/java/br/com/contmatic/empresav1/ITtest/DepartamentoIT.java
@@ -1,0 +1,11 @@
+package br.com.contmatic.empresav1.ITtest;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
+
+@RunWith(Suite.class)
+@SuiteClasses({})
+public class DepartamentoIT {
+
+}
